requirements.txt
streamlit
kafka-python
pandas
folium
requests
matplotlib
